--- Plutarch/Bool.hs-boot
@@ -1,7 +1,0 @@
-module Plutarch.Bool (PBool (..), pif) where
-
-import Plutarch.Internal.Term (S, Term)
-
-data PBool (s :: S) = PTrue | PFalse
-
-pif :: Term s PBool -> Term s a -> Term s a -> Term s a
