+++ Plutarch/Convert.hs
@@ -15,7 +15,8 @@ module Plutarch.Convert (
 ) where
 
 import GHC.Generics (Generic)
-import Plutarch.Bool (PBool (PFalse, PTrue), PEq, POrd, PPartialOrd)
+import Plutarch.Bool (PEq, POrd, PPartialOrd)
+import Plutarch.Builtin.Bool (PBool (PFalse, PTrue))
 import Plutarch.ByteString (PByteString)
 import Plutarch.Integer (PInteger)
 import Plutarch.Internal.Newtype (PlutusTypeNewtype)
