+++ Plutarch/Either.hs
@@ -26,7 +26,6 @@ module Plutarch.Either (
 import Data.Kind (Type)
 import GHC.Generics (Generic)
 import Plutarch.Bool (
-  PBool (PFalse, PTrue),
   PEq,
   POrd,
   PPartialOrd ((#<), (#<=)),
@@ -45,6 +44,7 @@ import Plutarch.Builtin (
   pfstBuiltin,
   psndBuiltin,
  )
+import Plutarch.Builtin.Bool (PBool (PFalse, PTrue))
 import Plutarch.Internal.Lift (
   DeriveDataPLiftable,
   PLiftable (
