+++ Plutarch/FFI.hs
@@ -31,8 +31,9 @@ import Generics.SOP.Type.Metadata (
   ConstructorName,
   DatatypeInfo (ADT, Newtype),
  )
-import Plutarch.Bool (PBool, PEq, (#==))
+import Plutarch.Bool (PEq, (#==))
 import Plutarch.Builtin (PData)
+import Plutarch.Builtin.Bool (PBool)
 import Plutarch.ByteString (PByteString)
 import Plutarch.Integer (PInteger)
 import Plutarch.Internal.Generic (PCode)
