+++ Plutarch/String.hs
@@ -16,7 +16,8 @@ import Data.String (IsString, fromString)
 import Data.Text (Text)
 import Data.Text qualified as Text
 import GHC.Generics (Generic)
-import Plutarch.Bool (PBool, PEq, (#&&), (#<=), (#==), (#||))
+import Plutarch.Bool (PEq, (#&&), (#<=), (#==), (#||))
+import Plutarch.Builtin.Bool (PBool)
 import Plutarch.ByteString (PByteString)
 import Plutarch.Integer (PInteger)
 import Plutarch.Internal.Lift (DeriveBuiltinPLiftable, PLiftable, PLifted (PLifted), pconstant)
